# Conflicts:
#	experiments/tunix-repro/base_config.py
#	src/jaxformers/attention_utils.py
#	src/jaxformers/masking_utils.py
#	src/jaxformers/models/huggingface/gemma3.py
#	src/jaxformers/scan_utils.py

+++ src/jaxformers/nn/linear.py
@@ -60,5 +60,5 @@ class Linear(eqx.Module):
             out_sharding=self.out_sharding,
         )
         if self.bias is not None:
-            y = y + self.bias[None, None, :]
+            y = y + self.bias.astype(x.dtype)[None, None, :]
         return y
